Rename modules in accordance with PEP8 and refactor package structure of db backends
