+++ aospy_synthetic/db/abstract_db.py
